+++ lib/src/internal/window_resize_listener.dart
@@ -20,8 +20,8 @@ class _WindowResizeListenerState extends State<WindowResizeListener>
 
   @override
   void initState() {
-    _lastSize = WidgetsBinding.instance!.window.physicalSize;
-    WidgetsBinding.instance!.addObserver(this);
+    _lastSize = WidgetsBinding.instance.window.physicalSize;
+    WidgetsBinding.instance.addObserver(this);
 
     widget.onResize(_lastSize);
 
@@ -30,13 +30,13 @@ class _WindowResizeListenerState extends State<WindowResizeListener>
 
   @override
   void dispose() {
-    WidgetsBinding.instance!.removeObserver(this);
+    WidgetsBinding.instance.removeObserver(this);
     super.dispose();
   }
 
   @override
   void didChangeMetrics() {
-    final winSize = WidgetsBinding.instance!.window.physicalSize;
+    final winSize = WidgetsBinding.instance.window.physicalSize;
 
     if (winSize != _lastSize) {
       widget.onResize(winSize);